refactor: #186 emby/jellyfin/plex 抽离为独立驱动子包
将 emby/jellyfin/plex 的 webhook 模型、数据提取、同步逻辑从
sync_service.py / models/sync.py / utils/data_util.py 抽离至各驱动
子包（app/services/{plex,emby,jellyfin}/），与 feiniu/fongmi 结构
对齐。

变更内容：
- 新增 plex/emby/jellyfin 子包：__init__.py、models.py、extractor.py、
  sync_service.py，分别承载 webhook 数据模型、报文提取、事件校验与分发
- SyncService.sync_plex_item / sync_emby_item / sync_jellyfin_item
  改为委托至各子包 sync_service，传入 self 作为 sync_svc 以兼容测试实例
- models/sync.py 与 utils/data_util.py 保留 re-export 以向后兼容
- 子包 sync_service 采用 late import 规避循环依赖
- 同步更新 test_sync_service*.py 中 9 处 patch 目标至新模块路径

修复：emby/jellyfin 子包 sync_service 误用模块级 sync_service 单例，
  改为使用传入的 sync_svc 参数（与 plex 子包一致）

测试：1893 passed（1 个 Windows mtime 分辨率导致的 flaky 用例另跑通过）
