+++ app/services/jellyfin/__init__.py
@@ -1,0 +1,5 @@
+from .extractor import extract_jellyfin_data
+from .models import JellyfinWebhookData
+from .sync_service import jellyfin_sync_service
+
+__all__ = ["extract_jellyfin_data", "JellyfinWebhookData", "jellyfin_sync_service"]
